+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - パフォーマンス検索を最適化
 - パフォーマンス残席数に、Chevreイベントのオファーごとの集計情報を利用するように調整
 - パフォーマンスからオファー情報を削除
+- onlineSalesStatus変更時にeventStatusも同時に変更するように調整
 
 ### Deprecated